+++ src/main/java/com/capstone/rentit/file/FileStorageException.java
@@ -1,0 +1,10 @@
+package com.capstone.rentit.file;
+
+public class FileStorageException extends RuntimeException {
+    public FileStorageException(String message) {
+        super(message);
+    }
+    public FileStorageException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
