+++ src/main/java/com/capstone/rentit/file/service/FileStorageService.java
@@ -1,4 +1,4 @@
-package com.capstone.rentit.file;
+package com.capstone.rentit.file.service;
 
 import org.springframework.web.multipart.MultipartFile;
