+++ src/main/java/com/capstone/rentit/rental/service/RentalService.java
@@ -1,6 +1,6 @@
 package com.capstone.rentit.rental.service;
 
-import com.capstone.rentit.file.FileStorageService;
+import com.capstone.rentit.file.service.FileStorageService;
 import com.capstone.rentit.member.dto.MemberDto;
 import com.capstone.rentit.rental.domain.Rental;
 import com.capstone.rentit.rental.dto.RentalDto;
